fix: added validation for the amazon link
