+++ .prettierrc
@@ -1,7 +1,6 @@
 {
 	"printWidth": 100,
 	"singleQuote": true,
-  "useTabs": false,
-  "tabWidth": 2,
+	"useTabs": false,
 	"trailingComma": "all"
 }
