+++ server/routes/users.js
@@ -28,13 +28,8 @@ const {
 } = require('../controller/messaging');
 const { handleError } = require('../helper/error');
 const log = require('../helper/logger');
-const {
-  redirectLogin,
-  redirectProfile,
-  verifyGoogleToken,
-  hashPassword,
-  createDefaultPassword,
-} = require('../helper/userHelper');
+const { verifyGoogleToken, hashPassword, createDefaultPassword } = require('../helper/user.helper');
+const { redirectLogin, redirectProfile } = require('./middleware/login.middleware');
 
 const UserRepository = require('../db/repository/UserRepository');
 const AgencyRepository = require('../db/repository/AgencyRepository');
